refactored init_img.c
